+++ apps/sv/src/main/scala/com/daml/network/sv/admin/api/client/commands/HttpSvAdminAppClient.scala
@@ -522,8 +522,8 @@ object HttpSvAdminAppClient {
       HttpResponse,
     ], http.GetDomainDataSnapshotResponse] =
       client.getDomainDataSnapshot(
-        body = definitions
-          .GetDomainDataSnapshotRequest(timestamp.toString, partyId.map(_.toProtoPrimitive)),
+        timestamp.toString,
+        partyId.map(_.toProtoPrimitive),
         headers = headers,
       )